+++ classes/Page.php
@@ -42,7 +42,6 @@ abstract class Page {
             case 'global':
                 $objHtmlPage = new HtmlPage();
                 $objHtmlPage->addExternalJS(APP_PATH_JS . 'base.js');
-                $objHtmlPage->addStylesheet('jquery-ui.min.css', 'screen,print');
                 $objHtmlPage->addStylesheet('style.css', 'screen,print');
                 $objHtmlPage->addStylesheet('home.css', 'screen,print');
                 $objHtmlPage->PrintHeader();